Версия 2.9.12.5 Доработана возможность приемки и перемещения кип макулатуры и материалов; переделана проверка доступности сервера онлайн; добавлена зависимость видимость операций (кнопок на основном окне) от роли пользователя

+++ gamma_mob/Models/MovementProduct.cs
@@ -12,7 +12,7 @@ namespace gamma_mob.Models
         //public string NomenclatureName { get; set; }
         //public string ShortNomenclatureName { get; set; }
         public Guid? PlaceZoneId { get; set; }
-        public int? ProductKindId { get; set; }
+        public byte? ProductKindId { get; set; }
         public int? CoefficientPackage { get; set; }
         public int? CoefficientPallet { get; set; }
         public int CollectedQuantityUnits { get; set; }
